Edit double to float.

+++ src/ru/job4j/converter/Converter.java
@@ -10,8 +10,8 @@ public class Converter {
         return  rls;
     }
     public static void main(String[] args) {
-        double euro = Converter.rubleToEuro(140);
-        double dollar = Converter.rubleToDollar(140);
+        float euro = Converter.rubleToEuro(140);
+        float dollar = Converter.rubleToDollar(140);
         System.out.println("140 ruble are " + euro + " euros. ");
         System.out.println("140 ruble are " + dollar + " dollars. ");
         //test Euro result
